# apiserver/src/config/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UUID": {
				"type": "string",
				"format": "uuid",
				"description": "Stringified UUIDv4.\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
			"ListIdDTO": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"email": {
				"type": "string",
				"format": "email",
				"description": "Stringified email.\nSee [Email Regex](https://emailregex.com/) and [RFC 5322](https://tools.ietf.org/html/rfc5322)",
				"pattern": "(([^<>()\\[\\]\\\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))"
			},
			"User": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"displayName": {
						"type": "string",
						"default": ""
					},
					"friends": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"lists": {
						"items": {
							"$ref": "#/components/schemas/List"
						},
						"type": "array"
					},
					"friendLists": {
						"items": {
							"$ref": "#/components/schemas/List"
						},
						"type": "array"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Boolean": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"List": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"title": {
						"type": "string",
						"default": ""
					},
					"closureDate": {
						"type": "string",
						"format": "date-time"
					},
					"owners": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"isShared": {
						"$ref": "#/components/schemas/Boolean"
					},
					"sharingCode": {
						"$ref": "#/components/schemas/UUID"
					},
					"grantedUsers": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time"
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"owners"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateListDTO": {
				"properties": {
					"title": {
						"type": "string",
						"default": ""
					},
					"closureDate": {
						"type": "string",
						"format": "date-time"
					},
					"owners": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"isShared": {
						"$ref": "#/components/schemas/Boolean"
					},
					"grantedUsers": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"owners"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ListDTO_": {
				"properties": {
					"sharingCode": {
						"$ref": "#/components/schemas/UUID"
					},
					"title": {
						"type": "string",
						"default": ""
					},
					"closureDate": {
						"type": "string",
						"format": "date-time"
					},
					"isShared": {
						"$ref": "#/components/schemas/Boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ListDTO": {
				"properties": {
					"sharingCode": {
						"$ref": "#/components/schemas/UUID"
					},
					"title": {
						"type": "string",
						"default": ""
					},
					"closureDate": {
						"type": "string",
						"format": "date-time"
					},
					"isShared": {
						"$ref": "#/components/schemas/Boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SelectKindList": {
				"description": "List Selection Kind",
				"enum": [
					"all",
					"owned",
					"granted"
				],
				"type": "string"
			},
			"UserIdDTO": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"MailAlreadyUsedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDTO": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"displayName": {
						"type": "string",
						"default": ""
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UserDTO_": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"displayName": {
						"type": "string",
						"default": ""
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserDTO": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/email"
					},
					"displayName": {
						"type": "string",
						"default": ""
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserNotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Giftlist API",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/lists": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListIdDTO"
								}
							}
						}
					}
				},
				"description": "Creates a new list.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateListDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ListDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Gets all user's lists data.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "select",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SelectKindList"
						}
					}
				]
			}
		},
		"/lists/{listId}": {
			"put": {
				"operationId": "Edit",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Edit a list.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of the list",
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"description": "data to edit a list",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ListDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Delete a list.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of the list",
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListDTO"
								}
							}
						}
					}
				},
				"description": "Sends back a list in the response based on its id.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of the list",
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/lists/{listId}/share": {
			"put": {
				"operationId": "Share",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Make a list to public.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of the list",
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/lists/{listId}/unshare": {
			"put": {
				"operationId": "Private",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Make a list private.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of the list",
						"in": "path",
						"name": "listId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/lists/invite/{sharingCode}": {
			"get": {
				"operationId": "AccessFromSharingCode",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Get a list from its sharing code.",
				"tags": [
					"List"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the sharing code of the list",
						"in": "path",
						"name": "sharingCode",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserIdDTO"
								}
							}
						}
					},
					"400": {
						"description": "If mail is already used",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MailAlreadyUsedError"
								}
							}
						}
					}
				},
				"description": "Create a new user during sign up. Even if users are authenticated and\ncreated by Auth0, we manage a user database to store preferences,\nfriends and much more.",
				"tags": [
					"User"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "data to create a user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Gets all user's data.",
				"tags": [
					"User"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"put": {
				"operationId": "Edit",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Edit a user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of user",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"description": "data to edit a user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UserDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					}
				},
				"description": "Delete logged user.",
				"tags": [
					"User"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of user",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					},
					"400": {
						"description": "If userId is incorrect",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserNotFoundError"
								}
							}
						}
					}
				},
				"description": "Gets logged user's data.",
				"tags": [
					"User"
				],
				"security": [
					{
						"auth0": []
					}
				],
				"parameters": [
					{
						"description": "the GUID of user",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}